requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pydantic[email]
openai
